refactor as I refine my understanding of web applications, specifically single page applications
