+++ app/index.css
@@ -15,7 +15,7 @@ body {
 .container, #app {
     width: 100%;
     height: 100%;
-    background-image: url('./images/pattern.svg');
+    background-image: url('./userinterface/resources/pattern.svg');
 }
 
 .home-container {
