+++ app/index.js
@@ -1,6 +1,6 @@
 let React = require('react');
 let ReactDOM = require('react-dom');
-let App = require('./components/layout/App');
+let App = require('./App');
 require('./index.css');
 
 ReactDOM.render(<App />,
